+++ service/reader/ReaderServiceImpl.cs
@@ -1,5 +1,7 @@
-﻿using Bookful.dao.reader;
+﻿using Bookful.dao.issuedBook;
+using Bookful.dao.reader;
 using Bookful.domain.dto;
+using Bookful.domain.exception;
 using Bookful.util.db;
 
 namespace Bookful.service.reader
@@ -7,10 +9,12 @@ namespace Bookful.service.reader
     public class ReaderServiceImpl : IReaderService
     {
         private readonly IReaderDao readerDao;
+        private readonly IIssuedBookDao issuedBookDao;
 
-        public ReaderServiceImpl(DBConnection connection)
+        public ReaderServiceImpl(IReaderDao readerDao, IIssuedBookDao issuedBookDao)
         {
-            this.readerDao = new ReaderDaoImpl(connection);
+            this.readerDao = readerDao;
+            this.issuedBookDao = issuedBookDao;
         }
 
         public bool AddReader(Reader reader)
@@ -20,7 +24,17 @@ namespace Bookful.service.reader
 
         public bool DeleteReaderById(int id)
         {
-            return readerDao.DeleteReaderById(id);
+            bool result = false;
+            if (HasReaderUnreturnedBooks(id))
+            {
+                throw new CommonException(domain.constant.Code.READER_NOT_RETURNED_BOOK, "У читателя есть невозвращенные книги");
+            }
+            else
+            {
+                result = readerDao.DeleteReaderById(id);
+            }
+
+            return result;
         }
 
         public List<Reader> GetAllReaders()
@@ -56,6 +70,21 @@ namespace Bookful.service.reader
                 return "Читатель не найден.";
             }
             return fullName;
+        }
+
+        public bool HasReaderUnreturnedBooks(int id)
+        {
+            bool result = false;
+            List<IssuedBook> issuedBooks = issuedBookDao.GetByReaderId(id);
+            foreach (var issuedBookItem in issuedBooks)
+            {
+                if (issuedBookItem.ReturnDate == null)
+                {
+                    result = true;
+                }
+            }
+
+            return result;
         }
 
         public List<Reader> SearchReaders(string searchText)
